+++ plugins/woocommerce-admin/client/analytics/components/report-filters/index.js
@@ -17,6 +17,7 @@ import { LOCALE } from '@woocommerce/wc-admin-settings';
  */
 import { recordEvent } from 'lib/tracks';
 import { Currency } from 'lib/currency-format';
+import { getCurrentDates, getDateParamsFromQuery, isoDateFormat } from 'lib/date';
 
 export default class ReportFilters extends Component {
 	constructor() {
@@ -64,6 +65,16 @@ export default class ReportFilters extends Component {
 
 	render() {
 		const { advancedFilters, filters, path, query, showDatePicker } = this.props;
+		const { period, compare, before, after } = getDateParamsFromQuery( query );
+		const { primary: primaryDate, secondary: secondaryDate } = getCurrentDates( query );
+		const dateQuery = {
+			period,
+			compare,
+			before,
+			after,
+			primaryDate,
+			secondaryDate,
+		};
 		return (
 			<Filters
 				query={ query }
@@ -76,6 +87,8 @@ export default class ReportFilters extends Component {
 				onDateSelect={ this.trackDateSelect }
 				onFilterSelect={ this.trackFilterSelect }
 				onAdvancedFilterAction={ this.trackAdvancedFilterAction }
+				dateQuery={ dateQuery }
+				isoDateFormat={ isoDateFormat }
 			/>
 		);
 	}