+++ plugins/woocommerce-admin/client/analytics/settings/default-date.js
@@ -3,6 +3,12 @@
  * External dependencies
  */
 import { parse, stringify } from 'qs';
+
+/**
+ * Internal dependencies
+ */
+import { getCurrentDates, getDateParamsFromQuery, isoDateFormat } from 'lib/date';
+
 /**
  * WooCommerce dependencies
  */
@@ -18,7 +24,24 @@ const DefaultDate = ( { value, onChange } ) => {
 		} );
 	};
 	const query = parse( value.replace( /&amp;/g, '&' ) );
-	return <DateRangeFilterPicker query={ query } onRangeSelect={ change } />;
+	const { period, compare, before, after } = getDateParamsFromQuery( query );
+	const { primary: primaryDate, secondary: secondaryDate } = getCurrentDates( query );
+	const dateQuery = {
+		period,
+		compare,
+		before,
+		after,
+		primaryDate,
+		secondaryDate,
+	};
+	return (
+		<DateRangeFilterPicker
+			query={ query }
+			onRangeSelect={ change }
+			dateQuery={ dateQuery }
+			isoDateFormat={ isoDateFormat }
+		/>
+	);
 };
 
 export default DefaultDate;