+++ plugins/woocommerce-admin/packages/components/src/filters/docs/example.js
@@ -9,6 +9,17 @@ import {
 	ReportFilters,
 	Section,
 } from '@woocommerce/components';
+import {
+	getDateParamsFromQuery,
+	getCurrentDates,
+	isoDateFormat,
+	loadLocaleData,
+} from '@woocommerce/date';
+
+/**
+ * External dependencies
+ */
+import { partialRight } from 'lodash';
 
 const ORDER_STATUSES = {
 	cancelled: 'Cancelled',
@@ -18,6 +29,25 @@ const ORDER_STATUSES = {
 	pending: 'Pending payment',
 	processing: 'Processing',
 	refunded: 'Refunded',
+};
+
+// Fetch locale from store settings and load for date functions.
+const localeSettings = {
+	userLocale: 'fr_FR',
+	weekdaysShort: [ 'dim', 'lun', 'mar', 'mer', 'jeu', 'ven', 'sam' ],
+};
+loadLocaleData( localeSettings );
+
+// Fetch store default date range and compose with date utility functions.
+const defaultDateRange = 'period=month&compare=previous_year';
+const storeGetDateParamsFromQuery = partialRight( getDateParamsFromQuery, defaultDateRange );
+const storeGetCurrentDates = partialRight( getCurrentDates, defaultDateRange );
+
+// Package date utilities for filter picker component.
+const storeDate = {
+	getDateParamsFromQuery: storeGetDateParamsFromQuery,
+	getCurrentDates: storeGetCurrentDates,
+	isoDateFormat,
 };
 
 const siteLocale = 'en_US';
@@ -180,7 +210,7 @@ export default () => (
 	<div>
 		<H>Date picker only</H>
 		<Section component={ false }>
-			<ReportFilters path={ path } query={ query } />
+			<ReportFilters path={ path } query={ query } storeDate={ storeDate } />
 		</Section>
 
 		<H>Date picker & more filters</H>
@@ -189,6 +219,7 @@ export default () => (
 				filters={ filters }
 				path={ path }
 				query={ query }
+				storeDate={ storeDate }
 			/>
 		</Section>